+++ history.md
@@ -1,4 +1,22 @@
 [2021/12/29]
+- Modify
+  + open.commons.spring.web.mvc.service.AbstractMvcService: DTO Class<?> 파라미터 제거
+    + selectMulti(SearchResultType, Function&lt;String[], Result&lt;List&lt;E&gt;&gt;&gt;, TripleFunction&lt;Integer, Integer, String[], Result&lt;List&lt;E&gt;&gt;&gt;, int, int, String[], Function&lt;E, D&gt;)
+    + selectMulti(SearchResultType, P, BiFunction&lt;P, String[], Result&lt;List&lt;E&gt;&gt;&gt;, QuadFunction&lt;P, Integer, Integer, String[], Result&lt;List&lt;E&gt;&gt;&gt;, int, int, String[], Function&lt;E, D&gt;)
+    + selectMulti(SearchResultType, P, Function&lt;P, Result&lt;List&lt;E&gt;&gt;&gt;, TripleFunction&lt;P, Integer, Integer, Result&lt;List&lt;E&gt;&gt;&gt;, int, int, Function&lt;E, D&gt;)
+    + selectMulti(SearchResultType, Supplier&lt;Result&lt;List&lt;E&gt;&gt;&gt;, BiFunction&lt;Integer, Integer, Result&lt;List&lt;E&gt;&gt;&gt;, int, int, Function&lt;E, D&gt;)
+    + selectMultiPagination(SearchResultType, P, Function&lt;P, Result&lt;Integer&gt;&gt;, BiFunction&lt;P, String[], Result&lt;List&lt;E&gt;&gt;&gt;, QuadFunction&lt;P, Integer, Integer, String[], Result&lt;List&lt;E&gt;&gt;&gt;, int, int, String[], Function&lt;E, D&gt;)
+    + selectMultiPagination(SearchResultType, P, Supplier&lt;Result&lt;Integer&gt;&gt;, Function&lt;P, Result&lt;List&lt;E&gt;&gt;&gt;, TripleFunction&lt;P, Integer, Integer, Result&lt;List&lt;E&gt;&gt;&gt;, int, int, Function&lt;E, D&gt;)
+    + selectMultiPagination(SearchResultType, Supplier&lt;Result&lt;Integer&gt;&gt;, Function&lt;String[], Result&lt;List&lt;E&gt;&gt;&gt;, TripleFunction&lt;Integer, Integer, String[], Result&lt;List&lt;E&gt;&gt;&gt;, int, int, String[], Function&lt;E, D&gt;)
+    + selectMultiPagination(SearchResultType, Supplier&lt;Result&lt;Integer&gt;&gt;, Supplier&lt;Result&lt;List&lt;E&gt;&gt;&gt;, BiFunction&lt;Integer, Integer, Result&lt;List&lt;E&gt;&gt;&gt;, int, int, Function&lt;E, D&gt;)
+  + open.commons.spring.web.mvc.service.IConvertingService: 변환 이후 Class<?> 파라미터 제거
+    + convertMultiPaginationResult(Result&lt;Page&lt;S&gt;&gt;, Function&lt;S, T&gt;)
+    + convertMultiResult(List&lt;S&gt;, Function&lt;S, T&gt;)
+    + convertMultiResult(Result&lt;List&lt;S&gt;&gt;, Function&lt;S, T&gt;)
+    + convertMultiResultAsStream(List&lt;S&gt;, Function&lt;S, T&gt;)
+    + convertSingleResult(Result&lt;S&gt;, Function&lt;S, T&gt;)
+
+[2021/12/29]
 - Add
   + open.commons.spring.web.mvc.service.AbstractMvcService
     + executePagination(Supplier&lt;Result&lt;List&lt;E&gt;&gt;&gt;, Supplier&lt;Result&lt;Integer&gt;&gt;, int, int, String[])